code/requirements.txt
alembic==1.7.4
boto3==1.18.60
botocore==1.21.60
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
firebase-admin==5.0.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.1.0
google-api-python-client==2.26.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.0
httplib2==0.20.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Mako==1.1.5
MarkupSafe==2.0.1
msgpack==1.0.2
numpy==1.21.3
packaging==21.0
Pillow==8.4.0
proto-plus==1.19.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.25
uritemplate==4.0.0
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.6.0
